pyproject.toml
[tool.poetry]
name = "theseus-alpha"
version = "0.1.0"
description = "LLM-driven trading system built on Hyperliquid"
authors = ["RAZVAN MIHAILESCU <RAZVAN.MIHAILESCU1996@GMAIL.COM>"]
readme = "README.md"
packages = [{ include = "src" }]
package-mode = true

[tool.poetry.dependencies]
python = "3.10.10"
hyperliquid-python-sdk = "^0.9.0" # Latest version
openai = "^1.0.0"
python-dotenv = "^1.0.0"
eth-utils = "^2.2.0"
eth-abi = "^4.2.1"
eth-account = "^0.10.0"
requests = "^2.31.0"
numpy = "^1.21.0"
pandas = "^1.3.0"
websocket-client = "^1.6.1"
msgpack = "^1.0.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
flake8 = "^4.0.0"
mypy = "^1.7.0"
jupyter = "^1.0.0"
ipython = "^8.0.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.poetry.scripts]
theseus = "theseus_alpha.cli.advisor:cli"
